basic save from request on the collector

+++ models/src/main/java/responses/SummonerResponse.java
@@ -2,8 +2,6 @@ package responses;
 
 import models.Summoner;
 
-import java.math.BigInteger;
-
 /**
  * Created by Maxime on 26/12/2015.
  */
@@ -12,7 +10,7 @@ public class SummonerResponse extends Response{
     public Integer id;
     public String name;
     public Integer profileIconId;
-    public BigInteger revisionDate;
+    public long revisionDate;
     public Integer summonerLevel;
 
     @Override
